+++ docs/changelog.md
@@ -2,9 +2,17 @@
 
 ## UNRELEASED
 
+## New features
+
+- Added filters on `random` manga
+
 ### Dependecies
 
 - Bump Pillow to v9.3.0
+
+### Breaking changes
+
+- `--search-filter` and `-sf` are no longer working as replaced with `--filter` and `-ft` option. It will be removed in v2.6.0
 
 ## v2.3.0